refactor: changing package of exception classes
