+++ src/main/java/it/gov/pagopa/fdr/repository/FdrPaymentRepository.java
@@ -16,8 +16,8 @@ import it.gov.pagopa.fdr.repository.common.Repository;
 import it.gov.pagopa.fdr.repository.common.RepositoryPagedResult;
 import it.gov.pagopa.fdr.repository.common.SortField;
 import it.gov.pagopa.fdr.repository.entity.payment.FdrPaymentEntity;
-import it.gov.pagopa.fdr.repository.exception.PersistenceFailureException;
-import it.gov.pagopa.fdr.repository.exception.TransactionRollbackException;
+import it.gov.pagopa.fdr.util.error.exception.persistence.PersistenceFailureException;
+import it.gov.pagopa.fdr.util.error.exception.persistence.TransactionRollbackException;
 import jakarta.enterprise.context.ApplicationScoped;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
