+++ src/main/java/it/gov/pagopa/fdr/service/FlowService.java
@@ -11,17 +11,17 @@ import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsPublishedR
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowCreatedResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
-import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
-import it.gov.pagopa.fdr.exception.AppException;
 import it.gov.pagopa.fdr.repository.FdrFlowRepository;
 import it.gov.pagopa.fdr.repository.FdrPaymentRepository;
 import it.gov.pagopa.fdr.repository.common.RepositoryPagedResult;
 import it.gov.pagopa.fdr.repository.entity.flow.FdrFlowEntity;
 import it.gov.pagopa.fdr.repository.enums.FlowStatusEnum;
-import it.gov.pagopa.fdr.repository.exception.PersistenceFailureException;
 import it.gov.pagopa.fdr.service.middleware.mapper.FlowMapper;
 import it.gov.pagopa.fdr.service.middleware.validator.SemanticValidator;
 import it.gov.pagopa.fdr.service.model.arguments.FindFlowsByFiltersArgs;
+import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
+import it.gov.pagopa.fdr.util.error.exception.common.AppException;
+import it.gov.pagopa.fdr.util.error.exception.persistence.PersistenceFailureException;
 import jakarta.enterprise.context.ApplicationScoped;
 import java.time.Instant;
 import java.util.concurrent.CompletableFuture;
