+++ src/main/java/it/gov/pagopa/fdr/service/middleware/validator/clause/PspValidator.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.fdr.service.middleware.validator.clause;
 
-import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
+import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdr.util.validator.ValidationArgs;
 import it.gov.pagopa.fdr.util.validator.ValidationResult;
 import it.gov.pagopa.fdr.util.validator.ValidationStep;
