+++ src/main/java/it/gov/pagopa/fdr/service/middleware/validator/clause/request/CreateFlowRequestValidator.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.fdr.service.middleware.validator.clause.request;
 
 import it.gov.pagopa.fdr.controller.model.flow.request.CreateFlowRequest;
-import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
+import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdr.util.validator.ValidationArgs;
 import it.gov.pagopa.fdr.util.validator.ValidationResult;
 import it.gov.pagopa.fdr.util.validator.ValidationStep;
