+++ src/main/java/it/gov/pagopa/fdr/util/error/exception/persistence/PersistenceFailureException.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.repository.exception;
+package it.gov.pagopa.fdr.util.error.exception.persistence;
 
 public class PersistenceFailureException extends Exception {
