+++ src/test/java/it/gov/pagopa/fdr/controller/psps/InternalPspResourceTest.java
@@ -28,10 +28,10 @@ import io.quarkus.test.junit.QuarkusTest;
 import it.gov.pagopa.fdr.controller.model.common.response.GenericResponse;
 import it.gov.pagopa.fdr.controller.model.error.ErrorResponse;
 import it.gov.pagopa.fdr.controller.model.flow.enums.SenderTypeEnum;
-import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdr.test.util.AzuriteResource;
 import it.gov.pagopa.fdr.test.util.MongoResource;
 import it.gov.pagopa.fdr.test.util.TestUtil;
+import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
